커뮤니티, 댓글 개발 (#94)
* 상품불러오기 개발 (#24)

* Feature/pmr select area (#26)

* [feat] getAreaListByUserId add

* [feat] getAreaListByUserId add

---------


* feat 유저의 지역목록 불러오기 쿼리수정

* [feat]User unRegister (#29)


* feat deleteAreaToUser (#30)


* [feat] deleteProduct add (#32)

* [feat] likeCount (#33)

* Feature/chatting (#38)

* feat 채팅개발

* Feat 채팅구현

* refactor 패키지 정리

* [feat] 채팅 보내기 개발

---------


* chore develop 머지 충돌 해결

* chore 병합충돌해결

* [feat] addAreaToUser add, [fix] 오타 수정 (#42)

* Feature/pmr addareatouser (#43)

* [feat] addAreaToUser add, [fix] 오타 수정

* [fix] getAreaListByUserId mdf, [feat] updateUserArea add

* updateproductBackUP (#51)

* [feat] getProductByUserId add (#52)

* [feat] increaseViewCount add (#56)

* [feat] refreshProduct add (#58)

* Feature/chatting read (#59)

* feat 위치 정보 정렬

* feat 로그인 추가

---------


* feat 유저개발

* [feat] getDefaultAreaByUserId add (#63)

* Feat 채팅목록 변경 (#64)

feat 상품올리기 비동기식에서 동기식으로 변경

* [feat] insertBoard add (#65)

* [feat]ProductShow server (#67)


* [feat] getBoardListByUserId add (#68)

* feat : 에러수정

* 럍: 병합해결

* fix : 상품올리기 에러 해결

* [feat] getBoardListByAreaId add, [fix] getBoardListByUserId query (#70)

* [feat] getDetailBoard add (#73)

* [feat] updateBoard add (#75)

* [feat] deleteBoard add (#78)

* Feat : 코드수정및 채팅로직 변경 (#81)

* [feat] insertComment add (#82)

* [feat] selectCommentById add (#84)

* [feat] getCommentByBoardId add (#86)

* [feat] increaseBoardViewCount add (#89)

* [fix] deleteBoard (#91)

* [feat] deleteComment add (#93)

---------

+++ src/main/java/com/example/carrot_market/board/service/BoardServiceImpl.java
@@ -5,6 +5,7 @@ import com.example.carrot_market.board.domain.model.Board;
 import com.example.carrot_market.board.domain.model.Comment;
 import com.example.carrot_market.board.dto.*;
 import com.example.carrot_market.core.error.CommonError;
+import com.example.carrot_market.user.db.UserMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -20,6 +21,8 @@ public class BoardServiceImpl implements BoardService {
 
     @Autowired
     private final BoardMapper boardMapper;
+    @Autowired
+    private final UserMapper userMapper;
 
     // 커뮤니티 글 작성
     @Override
@@ -65,11 +68,16 @@ public class BoardServiceImpl implements BoardService {
         return updateData;
     }
 
+    // 커뮤니티 조회수
     @Override
     public boolean increaseBoardViewCount(int boardId) {
-        return false;
+        Optional<Board> boardById = boardMapper.selectBoardById(boardId);
+        if(boardById.isEmpty()) throw new CommonError.Expected.ResourceNotFoundException("no exist board");
+
+        return boardMapper.increaseBoardViewCount(boardId);
     }
 
+    // 커뮤니티 좋아요
     @Override
     public void likeBoard(int boardId) {
 
@@ -83,6 +91,10 @@ public class BoardServiceImpl implements BoardService {
     // 선택한 커뮤니티 삭제
     @Override
     public void deleteBoard(int id) {
+        Optional<Board> boardById = boardMapper.selectBoardById(id);
+        if(boardById.isEmpty()) throw new CommonError.Expected.ResourceNotFoundException("no exist board");
+
+        boardMapper.deleteBoardWithComment(id);
         boardMapper.deleteBoard(id);
     }
 
@@ -93,7 +105,7 @@ public class BoardServiceImpl implements BoardService {
     }
 
     // 단일 댓글 조회
-    public Comment selectCommentById(int id) {
+    public Optional<Comment> selectCommentById(int id) {
         return boardMapper.selectCommentById(id);
     }
 
@@ -113,8 +125,11 @@ public class BoardServiceImpl implements BoardService {
     }
 
     @Override
-    public Comment deleteComment(int boardId, int commentId, int userId) {
-        return null;
+    public void deleteComment(int id) {
+        Optional<Comment> commentById = boardMapper.selectCommentById(id);
+        if(commentById.isEmpty()) throw new CommonError.Expected.ResourceNotFoundException("no exist comment");
+
+        boardMapper.deleteComment(id);
     }
 
     @Override
